Implemented a dedicated component for the triage table in FacilityHome
